fix(bezier-codemod): use react 18 for ink cli
