+++ .changeset/funny-ties-spark.md
@@ -1,0 +1,5 @@
+---
+"@channel.io/bezier-codemod": patch
+---
+
+Use React 18 for the Ink-based codemod CLI.
