--- .changeset/khaki-terms-march.md
@@ -1,5 +1,0 @@
----
-"@channel.io/bezier-codemod": patch
----
-
-Support non-interactive codemod execution with transform and path options.
